# Fix: EconomyInfo fixes & optimization