--- a/Scripts/Game/UI/HUD/OVT_EconomyInfo.c
+++ b/Scripts/Game/UI/HUD/OVT_EconomyInfo.c
@@ -95,30 +95,37 @@ class OVT_EconomyInfo : SCR_InfoDisplay {
 		}
 	}
 	
+	// FindOverride gets called a lot, so make it a tiny bit faster by having the pointer here
+	protected Managed m_foundComponent;
 	protected bool FindOverride(IEntity entity)
 	{
-		OVT_MainMenuContextOverrideComponent found = OVT_MainMenuContextOverrideComponent.Cast(entity.FindComponent(OVT_MainMenuContextOverrideComponent));
-		if(found) {
-			float dist = vector.Distance(entity.GetOrigin(),m_player.GetOrigin());
-			if(dist > found.m_fRange) return false;
-			if(m_fFoundRange == -1 || m_fFoundRange > dist)
+		m_foundComponent = entity.FindComponent(OVT_MainMenuContextOverrideComponent);
+		if(!m_foundComponent) return false;
+		
+		OVT_MainMenuContextOverrideComponent m_overrideComponent = OVT_MainMenuContextOverrideComponent.Cast(m_foundComponent);
+		if(!m_overrideComponent) return false;
+		
+		float dist = vector.Distance(entity.GetOrigin(),m_player.GetOrigin());
+		if(dist > m_overrideComponent.m_fRange) return false;
+		
+		if(m_fFoundRange == -1 || m_fFoundRange > dist)
+		{
+			bool got = true;
+			if(m_overrideComponent.m_bMustOwnBase)
 			{
-				bool got = true;
-				if(found.m_bMustOwnBase)
+				OVT_BaseData base = OVT_Global.GetOccupyingFaction().GetNearestBase(entity.GetOrigin());
+				if(!base || base.IsOccupyingFaction())
 				{
-					OVT_BaseData base = OVT_Global.GetOccupyingFaction().GetNearestBase(entity.GetOrigin());
-					if(!base || base.IsOccupyingFaction())
-					{
-						got = false;
-					}
+					got = false;
 				}
-				if(got)
-				{
-					m_FoundOverride = found;
-					m_fFoundRange = dist;
-				}
-			}			
+			}
+			if(got)
+			{
+				m_FoundOverride = m_overrideComponent;
+				m_fFoundRange = dist;
+			}
 		}
+		
 		return false;
 	}