In ``OVT_EconomyInfo``, ``m_fOverrideCounter`` is never reset, and override is updated every frame(?). Now it is set to update every second with a working timer. This is pretty significant because finding the override means checking the components of every entity within 50 meters, which is a really slow operation.

Average timings from the script profiler:

[Before](https://cdn.discordapp.com/attachments/512317720168103955/1250412050144821329/image.png?ex=666ad878&is=666986f8&hm=7efdea1bee4be2780792931db0fc994f08109646bdd6550e6c5b3c0d4aabcbb7&):
``OVT_EconomyInfo::UpdateOverride`` - 386.10us
``OVT_EconomyInfo::FindOverride`` - 265.50us

[After](https://cdn.discordapp.com/attachments/512317720168103955/1250429940034179184/image.png?ex=666ae921&is=666997a1&hm=ee13d055f2bebf0e59b7923ba886aded0e78c142977d90330a524b4cae53d940&):
``OVT_EconomyInfo::UpdateOverride`` - 5.50us
``OVT_EconomyInfo::FindOverride`` - 3.70us

One potential side-effect of this PR is that finding buildings for the UI override *might* sometimes be noticeably delayed, but I believe that this slight annoyance is worth it, and reading the code I believe this is what was originally intended anyway.

Edit:
Further optimized FindOverride by casting only when necessary and managing a single pointer instead of always creating a new one